Added the modules to the project
