--- modules/localStorage.js
@@ -1,3 +1,0 @@
-// // Retrieve saved books from localStorage, if any
-// const savedCollection = localStorage.getItem('bookCollection');
-// const bookCollection = savedCollection ? JSON.parse(savedCollection) : [];
